colocado diseño a todas las interfaces , agregado en ingresar con facebook , si no funciona debe colocarse el mismo codigo en login activity y arreglarlo
