+++ app/src/main/res/layout/activity_main.xml
@@ -7,9 +7,6 @@
     tools:context=".MainActivity"
     android:background="@drawable/bg_color"
     >
-<LinearLayout
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
@@ -22,7 +19,7 @@
             android:layout_height="wrap_content"
             android:gravity="center"
             android:minHeight="?actionBarSize"
-            android:padding="@dimen/appbar_padding"
+
             android:text="@string/app_name"
             android:textAppearance="@style/TextAppearance.Widget.AppCompat.Toolbar.Title" />
 
@@ -44,12 +41,11 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
         android:onClick="boton"
         app:backgroundTint="#FF5722"
         app:srcCompat="@android:drawable/ic_lock_power_off" />
 
 
-</LinearLayout>
+
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
